pyproject.toml
[project]
name = "inkbook-digest"
version = "0.1.0"
description = "Daily Readwise Reader → EPUB digest emailed to an Inkbook reader."
requires-python = "==3.12.*"
dependencies = [
    "httpx>=0.27",
    "ebooklib>=0.18",
    "apscheduler>=3.10",
    "pillow>=10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/digest"]
